+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## 0.9.0
 ### Features
-* better binary (options -f -c + read stdin + REPL)
+* better binary (options -f -c + read stdin + REPL + indent + catch exception)
+* indent API function
 ### Bugs
 * fix offset in tokenier
+* fix assign variables from variables (eg. function or variable aliases)
 
 ## 0.8.1
 ### Bugfix